Move some prettyprinting stuff from .Parse to .PrettyPrint
